Renaming and adding an api module to separate the external facing data structures from the internal ones.
