+++ apiDataStructures/README.md
@@ -1,0 +1,2 @@
+# API Data Structures
+Holds data structures that would be exposed to consumers of the REST services.  This is used internally, but could also be published as a library for JAVA REST implementations.
