+++ businessDataStructureBase/README.md
@@ -1,0 +1,2 @@
+# Business Data Structures
+Holds data structures that would be used by internal processing.
