+++ businessImplementationBase/README.md
@@ -1,0 +1,2 @@
+# Business Implementation
+Defines the implementation of the business interface.  This is where database acess, other data processing would occur that  is business specific.
