+++ businessImplementationBase/src/test/resources/META-INF/services/name.hampton.mike.IBusinessFunctionality
