+++ businessInterfaceBase/README.md
@@ -1,0 +1,2 @@
+# Business Interface
+Defines the business interface.  This is used to facilitate multiple business implementations.
