--- implementationBase/pom.xml
@@ -1,55 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>name.hampton.mike</groupId>
-  <artifactId>implementationBase</artifactId>
-	<properties>
-		<!-- configuration -->
-		<jdk.compiler.source>1.8</jdk.compiler.source>
-		<jdk.compiler.target>1.8</jdk.compiler.target>
-		<junit.version>4.11</junit.version>
-	</properties>
-  <packaging>jar</packaging>
-  <version>1.0-SNAPSHOT</version>
-  <name>implementationBase</name>
-  <url>http://maven.apache.org</url>
-  <dependencies>
-        <dependency>
-            <groupId>name.hampton.mike</groupId>
-            <artifactId>interfaceBase</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>name.hampton.mike</groupId>
-            <artifactId>interfaceBase</artifactId>
-            <version>${project.version}</version>
-            <type>test-jar</type>
-            <scope>test</scope>
-        </dependency>
-		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<version>${junit.version}</version>
-			<scope>test</scope>
-		</dependency>
-  </dependencies>
-
-	<build>
-		<pluginManagement>
-			<plugins>
-				<plugin>
-					<groupId>org.apache.maven.plugins</groupId>
-					<artifactId>maven-compiler-plugin</artifactId>
-					<version>3.2</version>
-
-					<configuration>
-						<source>${jdk.compiler.source}</source>
-						<target>${jdk.compiler.target}</target>
-						<encoding>UTF-8</encoding>
-					</configuration>
-
-				</plugin>
-			</plugins>
-		</pluginManagement>
-	</build>
-</project>
