--- interfaceBase/pom.xml
@@ -1,61 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<groupId>name.hampton.mike</groupId>
-	<artifactId>interfaceBase</artifactId>
-	<properties>
-		<!-- configuration -->
-		<jdk.compiler.source>1.8</jdk.compiler.source>
-		<jdk.compiler.target>1.8</jdk.compiler.target>
-		<junit.version>4.11</junit.version>
-		<!-- for the interface module we will not run the tests -->
-		<skipTests>true</skipTests>
-	</properties>
-
-	<packaging>jar</packaging>
-	<version>1.0-SNAPSHOT</version>
-	<name>interfaceBase</name>
-	<url>http://maven.apache.org</url>
-	<dependencies>
-		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<version>${junit.version}</version>
-			<scope>test</scope>
-		</dependency>
-	</dependencies>
-
-	<build>
-		<pluginManagement>
-			<plugins>
-				<plugin>
-					<groupId>org.apache.maven.plugins</groupId>
-					<artifactId>maven-compiler-plugin</artifactId>
-					<version>3.2</version>
-
-					<configuration>
-						<source>${jdk.compiler.source}</source>
-						<target>${jdk.compiler.target}</target>
-						<encoding>UTF-8</encoding>
-					</configuration>
-
-				</plugin>
-			</plugins>
-		</pluginManagement>
-		<plugins>
-			<!-- produce a jar which will include test classes and resources This is 
-				needed to allow the implementations to compile/run properly. -->
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-jar-plugin</artifactId>
-				<executions>
-					<execution>
-						<goals>
-							<goal>test-jar</goal>
-						</goals>
-					</execution>
-				</executions>
-			</plugin>
-		</plugins>
-	</build>
-</project>
