--- interfaceBase/src/main/java/name/hampton/mike/IBusinessFunctionality.java
@@ -1,23 +1,0 @@
-package name.hampton.mike;
-
-
-/**
- * The interface to the functionality.
- * 
- * Template to help with development of
- * a system.
- * 
- * @author mhampton
- *
- */
-public interface IBusinessFunctionality {
-	
-	public String getSimpleItem();
-	
-	public void setSimpleItem(String value);
-
-	public BusinessObject getComplexItem();
-	
-	public void setComplexItem(BusinessObject value);
-
-}
