--- interfaceBase/src/test/java/name/hampton/mike/IBusinessFunctionalityTest.java
@@ -1,78 +1,0 @@
-package name.hampton.mike;
-
-import java.util.Iterator;
-import java.util.ServiceLoader;
-
-import org.junit.Assert;
-import org.junit.Test;
-
-public class IBusinessFunctionalityTest {
-
-	@Test
-	public void testGetSimpleItem() {
-		final ServiceLoader<IBusinessFunctionality> serviceLoader = ServiceLoader.load(IBusinessFunctionality.class);
-		
-		Iterator<IBusinessFunctionality> impls = serviceLoader.iterator();
-		
-		if (!impls.hasNext()) Assert.fail("No implementations found to test!");
-		while (impls.hasNext()) {
-			IBusinessFunctionality instance = null;
-			instance = impls.next();
-			String actual = instance.getSimpleItem();
-			
-			Assert.assertEquals(String.format("Implementation is %s", instance.getClass().getName()), "simple", actual);
-		}		
-	}
-
-	@Test
-	public void testSetSimpleItem() {
-		final ServiceLoader<IBusinessFunctionality> serviceLoader = ServiceLoader.load(IBusinessFunctionality.class);
-		
-		Iterator<IBusinessFunctionality> impls = serviceLoader.iterator();
-		
-		if (!impls.hasNext()) Assert.fail("No implementations found to test!");
-		while (impls.hasNext()) {
-			IBusinessFunctionality instance = null;
-			instance = impls.next();
-			instance.setSimpleItem("testSet");
-			String actual = instance.getSimpleItem();
-			
-			Assert.assertEquals(String.format("Implementation is %s", instance.getClass().getName()), "testSet", actual);
-		}		
-	}
-
-	@Test
-	public void testGetComplexItem() {
-		final ServiceLoader<IBusinessFunctionality> serviceLoader = ServiceLoader.load(IBusinessFunctionality.class);
-		
-		Iterator<IBusinessFunctionality> impls = serviceLoader.iterator();
-		
-		if (!impls.hasNext()) Assert.fail("No implementations found to test!");
-		while (impls.hasNext()) {
-			IBusinessFunctionality instance = null;
-			instance = impls.next();
-			BusinessObject actual = instance.getComplexItem();
-			
-			Assert.assertEquals(String.format("Implementation is %s", instance.getClass().getName()), null, actual.getText());
-		}		
-	}
-
-	@Test
-	public void testSetComplexItem() {
-		final ServiceLoader<IBusinessFunctionality> serviceLoader = ServiceLoader.load(IBusinessFunctionality.class);
-		
-		Iterator<IBusinessFunctionality> impls = serviceLoader.iterator();
-		
-		if (!impls.hasNext()) Assert.fail("No implementations found to test!");
-		while (impls.hasNext()) {
-			IBusinessFunctionality instance = null;
-			instance = impls.next();
-			BusinessObject value = new BusinessObject();
-			value.setText("testSet");
-			instance.setComplexItem(value);
-			BusinessObject actual = instance.getComplexItem();
-			
-			Assert.assertEquals(String.format("Implementation is %s", instance.getClass().getName()), value.getText(), actual.getText());
-		}		
-	}
-}
