+++ pom.xml
@@ -2,47 +2,49 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
                              http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
+  <modelVersion>4.0.0</modelVersion>
 
-    <groupId>name.hampton.mike</groupId>
-    <artifactId>restParent</artifactId>
-    <properties>
-        <jdk.compiler.source>1.7</jdk.compiler.source>
-        <jdk.compiler.target>1.7</jdk.compiler.target>
-        <junit.version>4.11</junit.version>
-    </properties>
+  <groupId>name.hampton.mike</groupId>
+  <artifactId>restParent</artifactId>
+  <properties>
+    <jdk.compiler.source>1.7</jdk.compiler.source>
+    <jdk.compiler.target>1.7</jdk.compiler.target>
+    <junit.version>4.11</junit.version>
+  </properties>
 
-    <packaging>pom</packaging>
-    <version>1.0</version>
-    <name>Rest Parent Project</name>
+  <packaging>pom</packaging>
+  <version>1.0</version>
+  <name>Rest Parent Project</name>
 
-    <modules>
-        <module>interfaceBase</module>
-        <module>implementationBase</module>
-        <module>restInterface</module>
-    </modules>
+  <modules>
+    <module>businessInterfaceBase</module>
+    <module>businessImplementationBase</module>
+    <module>businessDataStructureBase</module>
+    <module>restInterface</module>
+    <module>apiDataStructures</module>
+  </modules>
 
-    <build>
-        <pluginManagement>
-            <plugins>
-                <plugin>
-                    <groupId>org.apache.maven.plugins</groupId>
-                    <artifactId>maven-compiler-plugin</artifactId>
-                    <configuration>
-                        <source>${jdk.compiler.source}</source>
-                        <target>${jdk.compiler.target}</target>
-                    </configuration>
-                </plugin>
-            </plugins>
-        </pluginManagement>
-    </build>
+  <build>
+    <pluginManagement>
+      <plugins>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-compiler-plugin</artifactId>
+          <configuration>
+            <source>${jdk.compiler.source}</source>
+            <target>${jdk.compiler.target}</target>
+          </configuration>
+        </plugin>
+      </plugins>
+    </pluginManagement>
+  </build>
 
-    <dependencies>
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>${junit.version}</version>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
+  <dependencies>
+    <dependency>
+      <groupId>junit</groupId>
+      <artifactId>junit</artifactId>
+      <version>${junit.version}</version>
+      <scope>test</scope>
+    </dependency>
+  </dependencies>
 </project>
