+++ restInterface/pom.xml
@@ -1,96 +1,102 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
-    <modelVersion>4.0.0</modelVersion>
+  <modelVersion>4.0.0</modelVersion>
 
-    <groupId>name.hampton.mike</groupId>
-    <artifactId>restInterface</artifactId>
-	<properties>
-		<!-- configuration -->
-		<jdk.compiler.source>1.8</jdk.compiler.source>
-		<jdk.compiler.target>1.8</jdk.compiler.target>
-		<junit.version>4.11</junit.version>
+  <groupId>name.hampton.mike</groupId>
+  <artifactId>restInterface</artifactId>
+  <properties>
+    <!-- configuration -->
+    <jdk.compiler.source>1.8</jdk.compiler.source>
+    <jdk.compiler.target>1.8</jdk.compiler.target>
+    <junit.version>4.11</junit.version>
 
-        <jersey.version>2.17</jersey.version>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-	</properties>
+    <jersey.version>2.17</jersey.version>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+  </properties>
 
-    <packaging>jar</packaging>
-    <version>1.0-SNAPSHOT</version>
-    <name>restInterface</name>
+  <packaging>jar</packaging>
+  <version>1.0-SNAPSHOT</version>
+  <name>restInterface</name>
 
-    <dependencyManagement>
-        <dependencies>
-            <dependency>
-                <groupId>org.glassfish.jersey</groupId>
-                <artifactId>jersey-bom</artifactId>
-                <version>${jersey.version}</version>
-                <type>pom</type>
-                <scope>import</scope>
-            </dependency>
-        </dependencies>
-    </dependencyManagement>
-
+  <dependencyManagement>
     <dependencies>
-        <dependency>
-            <groupId>org.glassfish.jersey.containers</groupId>
-            <artifactId>jersey-container-grizzly2-http</artifactId>
-        </dependency>
-        <!-- uncomment this to get JSON support:
-        -->
-         <dependency>
-            <groupId>org.glassfish.jersey.media</groupId>
-            <artifactId>jersey-media-moxy</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>name.hampton.mike</groupId>
-            <artifactId>interfaceBase</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>name.hampton.mike</groupId>
-            <artifactId>implementationBase</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-			<version>${junit.version}</version>
-			<scope>test</scope>
-        </dependency>
+      <dependency>
+        <groupId>org.glassfish.jersey</groupId>
+        <artifactId>jersey-bom</artifactId>
+        <version>${jersey.version}</version>
+        <type>pom</type>
+        <scope>import</scope>
+      </dependency>
     </dependencies>
+  </dependencyManagement>
 
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>2.5.1</version>
-                <inherited>true</inherited>
-                <configuration>
-						<source>${jdk.compiler.source}</source>
-						<target>${jdk.compiler.target}</target>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>org.codehaus.mojo</groupId>
-                <artifactId>exec-maven-plugin</artifactId>
-                <version>1.2.1</version>
-                <executions>
-                    <execution>
-                        <goals>
-                            <goal>java</goal>
-                        </goals>
-                    </execution>
-                </executions>
-                <configuration>
-                    <mainClass>name.hampton.mike.Main</mainClass>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
+  <dependencies>
+    <dependency>
+      <groupId>org.glassfish.jersey.containers</groupId>
+      <artifactId>jersey-container-grizzly2-http</artifactId>
+    </dependency>
+    <!-- uncomment this to get JSON support:
+    -->
+    <dependency>
+      <groupId>org.glassfish.jersey.media</groupId>
+      <artifactId>jersey-media-moxy</artifactId>
+    </dependency>
+
+    <dependency>
+      <groupId>name.hampton.mike</groupId>
+      <artifactId>businessInterfaceBase</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+
+    <dependency>
+      <groupId>name.hampton.mike</groupId>
+      <artifactId>businessImplementationBase</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+
+    <dependency>
+      <groupId>name.hampton.mike</groupId>
+      <artifactId>apiDataStructures</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+
+    <dependency>
+      <groupId>junit</groupId>
+      <artifactId>junit</artifactId>
+      <version>${junit.version}</version>
+      <scope>test</scope>
+    </dependency>
+  </dependencies>
+
+  <build>
+    <plugins>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-compiler-plugin</artifactId>
+        <version>2.5.1</version>
+        <inherited>true</inherited>
+        <configuration>
+          <source>${jdk.compiler.source}</source>
+          <target>${jdk.compiler.target}</target>
+        </configuration>
+      </plugin>
+      <plugin>
+        <groupId>org.codehaus.mojo</groupId>
+        <artifactId>exec-maven-plugin</artifactId>
+        <version>1.2.1</version>
+        <executions>
+          <execution>
+            <goals>
+              <goal>java</goal>
+            </goals>
+          </execution>
+        </executions>
+        <configuration>
+          <mainClass>name.hampton.mike.Main</mainClass>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
 
 </project>
