+++ restInterface/src/main/java/name/hampton/mike/MyResource2.java
@@ -1,5 +1,8 @@
 package name.hampton.mike;
 
+import name.hampton.mike.internal.BusinessObject;
+import name.hampton.mike.sdk.V1BusinessObject;
+
 import java.util.Iterator;
 import java.util.ServiceLoader;
 import java.util.logging.Logger;
@@ -20,21 +23,12 @@ public class MyResource2 {
   private Logger logger = Logger.getLogger(this.getClass().getName());
 
   /**
-   * Method handling HTTP GET requests. The returned object will be sent
-   * to the client as "text/plain" media type.
+   * Load the first implementation found.
    *
-   * @return String that will be returned as a text/plain response.
-   * @throws ImplementationNotFoundException - if the ServiceLoader cannot find the implementation.
+   * @return an IBusinessFunctionality instance
+   * @throws ImplementationNotFoundException - if no IBusinessFunctionality is configured.
+   * @see ServiceLoader#load(java.lang.Class)
    */
-  @GET
-  @Produces(MediaType.TEXT_PLAIN)
-  public String getIt(@PathParam("resourceId") String resourceId) throws ImplementationNotFoundException {
-    IBusinessFunctionality instance = getiBusinessFunctionality();
-    String actual = instance.getSimpleItem();
-
-    return String.format("resourceId=%s, item=%s", resourceId, actual);
-  }
-
   private IBusinessFunctionality getiBusinessFunctionality() throws ImplementationNotFoundException {
     final ServiceLoader<IBusinessFunctionality> serviceLoader = ServiceLoader.load(IBusinessFunctionality.class);
 
@@ -50,6 +44,21 @@ public class MyResource2 {
     return instance;
   }
 
+  /**
+   * Method handling HTTP GET requests. The returned object will be sent
+   * to the client as "text/plain" media type.
+   *
+   * @return String that will be returned as a text/plain response.
+   * @throws ImplementationNotFoundException - if the ServiceLoader cannot find the implementation.
+   */
+  @GET
+  @Produces(MediaType.TEXT_PLAIN)
+  public String getIt(@PathParam("resourceId") String resourceId) throws ImplementationNotFoundException {
+    IBusinessFunctionality instance = getiBusinessFunctionality();
+    String actual = instance.getSimpleItem();
+
+    return String.format("resourceId=%s, item=%s", resourceId, actual);
+  }
 
   /**
    * Method handling HTTP GET requests. The returned object will be sent
@@ -59,24 +68,28 @@ public class MyResource2 {
    * @throws ImplementationNotFoundException - if the ServiceLoader cannot find the implementation.
    */
   @GET
-  @Produces("application/json")
-  public BusinessObject getItAsJson(@PathParam("resourceId") String resourceId) throws ImplementationNotFoundException {
+  @Produces(MediaType.APPLICATION_JSON)
+  public V1BusinessObject getItAsJson(@PathParam("resourceId") String resourceId) throws ImplementationNotFoundException {
     IBusinessFunctionality instance = getiBusinessFunctionality();
     BusinessObject actual = instance.getComplexItem();
     actual.setText(resourceId);
 
-    return actual;
+    return BusinessObjectToSdkObject.convert(actual);
   }
 
   @POST
   @Path("/transformStuff")
   @Produces(MediaType.APPLICATION_JSON)
   @Consumes(MediaType.APPLICATION_JSON)
-  public BusinessObject transformStuff(@PathParam("resourceId") String resourceId, BusinessObject actual) {
+  public V1BusinessObject transformStuff(@PathParam("resourceId") String resourceId, V1BusinessObject actualV1) {
+    if (actualV1!=null) {
+      BusinessObject actual = BusinessObjectToSdkObject.convert(actualV1);
 
-    actual.setText(String.format("transformed! resourceId is %s, origtext is %s", resourceId, actual.getText()));
-    actual.setNumber(actual.getNumber() + 1);
+      actual.setText(String.format("transformed! resourceId is %s, origtext is %s", resourceId, actual.getText()));
+      actual.setNumber(actual.getNumber() + 1);
 
-    return actual;
+      return BusinessObjectToSdkObject.convert(actual);
+    }
+    return null;
   }
 }
